Recover stack-local widths and file mode globals

+++ include/game/military/TCivUnit.h
@@ -15,7 +15,7 @@ public:
   virtual ~TCivUnit() override;                    // slot 0x01 (scalar deleting destructor)
   virtual void WriteTo(TStream* stream) override;  // slot 0x05 0x5c2b40
   virtual void ReadFrom(TStream* stream) override; // slot 0x06 0x5c2b10
-  virtual void MoveTo(int nTileIndex) override;    // slot 0x0a 0x5c2b70
+  virtual void MoveTo(short nTileIndex) override;  // slot 0x0a 0x5c2b70
   virtual void ContinueOrders() override;          // slot 0x0b 0x5c2a90
   virtual void DetachUnitOrderFromOwnerAndReset() override;      // slot 0x0c 0x5c2c40
   virtual void SetOrders(UnitOrder order, int payload) override; // slot 0x0d 0x5c29f0
